fix: support negative slice on eager frame

+++ py-polars/src/dataframe.rs
@@ -1098,10 +1098,10 @@ impl PyDataFrame {
         Ok(())
     }
 
-    pub fn slice(&self, offset: usize, length: Option<usize>) -> Self {
+    pub fn slice(&self, offset: i64, length: Option<usize>) -> Self {
         let df = self
             .df
-            .slice(offset as i64, length.unwrap_or_else(|| self.df.height()));
+            .slice(offset, length.unwrap_or_else(|| self.df.height()));
         df.into()
     }
